+++ src/permissions/modDefinitionPermissionsHelpers.ts
